pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.115.11",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "joblib>=1.4.2",
    "numpy>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.6.1",
    "uvicorn>=0.34.0",
]
